requirements.txt
requests
tweepy
python-dotenv
feedparser